feat:
Due to the deep modifications related to names and packet format,
this task cover a large part of the codebase and involves several changes:

- the library provides a name data structure (hicn_name_t ), which is composed
  of a name prefix (hicn_name_prefix_t) and a name suffix (hicn_name_suffix_t),
  and it has been extended to provide all support functions required for name
  manipulation, including common prefix computation, as required for the Longest
  Prefix Match (LPM)in the forwarder, in addition to Exact Prefix Match (EPM).
- all code has been rewritten to use this data structure instead of having for
  instance the forwarder define its own name class (used to be Name and NameBitVector)
  the code has been refactored to minimize name allocations and copies, one remaining
  aspect is the difference of name storage between PIT and CS entries (respectively
  in the PIT entry, and in the message buffer), which causes the packet cache
  index to be updated when a PIT entry is converted into a CS entry. By storing
  the name in the PIT/CS entry everytime, we might save on this operation).
- hicn-light FIB has been rewritten : code has been refactored and should now be
  shorter and documented; unit tests have been drafted but more would be required
  to cover all cases and match the algorithms to add/remove nodes, as specified in the doc.
  all protocol details and hICN header formats are now abstracted by the library
  for the forwarder (and thus header.h and  protocols/*.h have been removed from
  public includes, and replaced by packet.h providing protocol agnostic packet
  level functions, completely replacing the compat.h header that used to provide
  similar functions.
- this works by exposing a opaque buffer to the application (a kind of socket buffer)
  which is used by the lib to cache the packet format and offsets of the different
  layers in the buffer and provider efficient operations (the packet format is
  either defined for packet construction, or guessed at ingress, and this structure
  is updated accordingly only once).

+++ lib/includes/hicn/util/vector.h
@@ -209,7 +209,7 @@ static inline int
 _vector_get (void *vector, off_t pos, size_t elt_size, void *elt)
 {
   vector_hdr_t *vh = vector_hdr (vector);
-  if (pos >= vh->alloc_size)
+  if ((size_t) pos >= vh->alloc_size)
     return -1;
 
   memcpy (elt, (uint8_t *) vector + pos * elt_size, elt_size);
@@ -228,7 +228,7 @@ _vector_get (void *vector, off_t pos, size_t elt_size, void *elt)
 static inline bool
 _vector_contains (void *vector, size_t elt_size, void *elt)
 {
-  for (int i = 0; i < vector_hdr (vector)->cur_size; i++)
+  for (size_t i = 0; i < vector_hdr (vector)->cur_size; i++)
     {
       if (memcmp ((uint8_t *) vector + i * elt_size, elt, elt_size) == 0)
 	return true;
@@ -251,7 +251,7 @@ static inline int
 _vector_remove_at (void **vector_ptr, size_t elt_size, off_t pos)
 {
   vector_hdr_t *vh = vector_hdr (*vector_ptr);
-  if (pos >= vh->cur_size)
+  if ((size_t) pos >= vh->cur_size)
     return -1;
 
   // Shift backward by one position all the elements after the one specified
@@ -342,7 +342,7 @@ _vector_remove_at (void **vector_ptr, size_t elt_size, off_t pos)
 
 #define vector_at(vector, pos)                                                \
   ({                                                                          \
-    assert (pos < vector_hdr (vector)->cur_size);                             \
+    assert ((size_t) pos < vector_hdr (vector)->cur_size);                    \
     (vector)[(pos)];                                                          \
   })
